fixing ratelimit on api

+++ config/api.php
@@ -1,5 +1,5 @@
 <?php
 
 return [
-    'rate_limit' => env('API_RATE_LIMIT', 60),
+    'rate_limit' => (int) env('API_RATE_LIMIT', 60),
 ];
